move sections into own module (again). add back confirm write button
